NoteVault/requirements.txt
asgiref==3.7.2
Django==4.1.13
django-cors-headers==4.3.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djongo==1.3.6
dnspython==2.4.2
Pillow==10.1.0
PyJWT==2.8.0
pymongo==4.6.0
pytz==2023.3.post1
sqlparse==0.2.4
tzdata==2023.3
